Clean up servlets
