+++ server/pom.xml
@@ -126,7 +126,7 @@
         <artifactId>exec-maven-plugin</artifactId>
         <version>1.2.1</version>
         <configuration>
-            <mainClass>com.google.sps.utils.StatesCollectionBuilder</mainClass>
+            <mainClass>com.google.sps.utils.StateCollectionBuilder</mainClass>
         </configuration>
       </plugin>
     </plugins>
